NXDRIVE-279: Refactor LocalClient
