+++ docs/changes/4.2.1.md
@@ -4,6 +4,7 @@ Release date: `2019-xx-xx`
 
 ## Core
 
+- [NXDRIVE-279](https://jira.nuxeo.com/browse/NXDRIVE-279): Refactor `LocalClient`
 - [NXDRIVE-1838](https://jira.nuxeo.com/browse/NXDRIVE-1838): Allow for one time synchronization of a file (direct upload)
 - [NXDRIVE-1851](https://jira.nuxeo.com/browse/NXDRIVE-1851): The config parser does not handle float values
 - [NXDRIVE-1859](https://jira.nuxeo.com/browse/NXDRIVE-1859): [Windows] Fix the special file check for folder icon
@@ -35,6 +36,7 @@ Release date: `2019-xx-xx`
 - Upgraded `sentry-sdk` from 0.11.2 to 0.12.2
 - Upgraded `tld` from 0.9.3 to 0.9.6
 - Upgraded `urllib3` from 1.25.3 to 1.25.6
+- Removed `send2trash`
 
 ## Technical Changes
 
@@ -44,6 +46,10 @@ Release date: `2019-xx-xx`
 - Added `local_state` keyword argument to `EngineDAO.insert_local_state()`, defaults to `created` to not change the behavior.
 - Added `check` keyword argument to `LocalClient.get_info()`, defaults to `True` to not change the behavior.
 - Added `Remote.direct_upload()`
+- Added client/local/base.py
+- Added client/local/darwin.py
+- Added client/local/linux.py
+- Added client/local/windows.py
 - Added gui/folders_dialog.py
 - Added gui/folders_loader.py
 - Added gui/folders_model.py
@@ -53,3 +59,4 @@ Release date: `2019-xx-xx`
 - Removed `DirectEdit.url`
 - Removed `FiltersDialog`. Use `DocumentsDialog` instead.
 - Removed `FolderTreeview`. Use `DocumentTreeView` instead.
+- Removed client/local_client.py
