+++ tests/old_functional/__init__.py
@@ -10,7 +10,7 @@ import nuxeo.constants
 import nuxeo.operations
 from nuxeo.exceptions import HTTPError
 
-from nxdrive.client.local_client import LocalClient
+from nxdrive.client.local import LocalClient
 from nxdrive.client.remote_client import Remote
 from nxdrive.objects import NuxeoDocumentInfo, RemoteFileInfo
 from nxdrive.options import Options
